requirements/requirements.txt
tkinter
pyserial
threading
json
requests
RPi.GPIO
